+++ tidy3d/components/grid/grid_spec.py
@@ -10,7 +10,6 @@ import pydantic.v1 as pd
 
 from tidy3d.components.base import Tidy3dBaseModel, cached_property, skip_if_fields_missing
 from tidy3d.components.geometry.base import Box, ClipOperation
-from tidy3d.components.lumped_element import LumpedElementType
 from tidy3d.components.source.utils import SourceType
 from tidy3d.components.structure import MeshOverrideStructure, Structure, StructureType
 from tidy3d.components.types import (
@@ -27,6 +26,7 @@ from tidy3d.components.types import (
 from tidy3d.constants import C_0, MICROMETER, dp_eps, inf
 from tidy3d.exceptions import SetupError
 from tidy3d.log import log
+from tidy3d.microwave.lumped_element import LumpedElementType
 
 from .corner_finder import CornerFinderSpec
 from .grid import Coords, Coords1D, Grid
