+++ tidy3d/microwave/ports/base_lumped.py
@@ -10,10 +10,10 @@ import pydantic.v1 as pd
 from tidy3d.components.base import cached_property
 from tidy3d.components.geometry.utils_2d import snap_coordinate_to_grid
 from tidy3d.components.grid.grid import Grid, YeeGrid
-from tidy3d.components.lumped_element import LumpedElementType
 from tidy3d.components.monitor import FieldMonitor
 from tidy3d.components.types import Complex, Coordinate, FreqArray
 from tidy3d.constants import OHM
+from tidy3d.microwave.lumped_element import LumpedElementType
 
 from .base_terminal import AbstractTerminalPort
