+++ tidy3d/microwave/ports/rectangular_lumped.py
@@ -19,13 +19,13 @@ from tidy3d.components.geometry.utils import (
 )
 from tidy3d.components.geometry.utils_2d import increment_float
 from tidy3d.components.grid.grid import Grid, YeeGrid
-from tidy3d.components.lumped_element import LinearLumpedElement, LumpedResistor, RLCNetwork
 from tidy3d.components.monitor import FieldMonitor
 from tidy3d.components.source.current import UniformCurrentSource
 from tidy3d.components.source.time import GaussianPulse
 from tidy3d.components.types import Axis, FreqArray, LumpDistType
 from tidy3d.components.validators import assert_line_or_plane
 from tidy3d.exceptions import SetupError, ValidationError
+from tidy3d.microwave.lumped_element import LinearLumpedElement, LumpedResistor, RLCNetwork
 from tidy3d.plugins.microwave import CurrentIntegralAxisAligned, VoltageIntegralAxisAligned
 
 from .base_lumped import AbstractLumpedPort
